feat: adição do TCLE
